Split off DevRandom into it's own project.
As it stands this code is a couple of days work to add a java.util.Random which
could use an operating system's random device, initially targeting the Android
operating system for use with Hashwords. As you can probably see it has startedc
to expand into other operating systems already.

The initial driving force behind DevRandom were the potentially insecure random
numbers on Android 4.3 and older.

https://android-developers.blogspot.com/2013/08/some-securerandom-thoughts.html

Judging by that blog post the initial code was written early 2014. I have a bac
up of the filesystem somewhere so I may be to find the actual dates, but not
today. I'm splitting it off into it's own repository so I can push on with
development.